all/net: fix GSO send fallback, byte order, and alignment
Fixes review feedback on PR #39:

#21 (High) GSO send lacks runtime fallback:
  - When sendToGSO fails, fallback to regular UDP send instead of returning error
  - Changed err != nil check to err == nil with fallback path

#22 (Medium) Zig GSO uses wrong byte order:
  - Changed .little to .native for host byte order
  - Kernel expects host byte order for UDP_SEGMENT

#23 (Medium) Go GSO cmsg data pointer alignment:
  - Improved comment explaining alignment
  - Use uintptr cast for proper pointer arithmetic

+++ go/pkg/net/peer_linux.go
@@ -40,8 +40,8 @@ func (u *UDP) sendToGSO(data []byte, addr *net.UDPAddr) (int, error) {
 		}
 
 		// Write segment size (uint16) into cmsg data area
-		// unix.CmsgData returns unsafe.Pointer to the data portion
-		dataPtr := unsafe.Pointer(uintptr(unsafe.Pointer(cmsg)) + unix.SizeofCmsghdr)
+		// cmsg data starts after the cmsghdr with proper alignment
+		dataPtr := unsafe.Pointer(uintptr(unsafe.Pointer(cmsg)) + uintptr(unix.SizeofCmsghdr))
 		dataBuf := (*[2]byte)(dataPtr)
 		binary.NativeEndian.PutUint16(dataBuf[:], uint16(DefaultGSOSegment))
